@@ -17,15 +17,13 @@ classifiers =
 [options]
 #package_data = {'oxdna_analysis_tools': ['src.oxDNA_analysis_tools/cython_utils/*.pyx']}
 include_package_data = True
-package_dir =
-    = src
+package_dir = = src
 packages = find:
-python_requires = >=3.8
+python_requires = >=3.9
 install_requires =
 		 numpy > 1.16
 		 scikit-learn > 0.21.2
 		 matplotlib > 3.0.0
-         biopython > 1.73
 
 [options.packages.find]
 where = src
