@@ -1,11 +1,9 @@
 import numpy as np
 import itertools
-from math import sqrt
-import sys
 import copy
-from typing import List
+from typing import List, Dict
 
-BASE_SHIFT = 1.13
+BASE_SHIFT = 3.4
 COM_SHIFT = 0.5
 FROM_OXDNA_TO_ANGSTROM = 8.518
 FROM_ANGSTROM_TO_OXDNA = 1. / FROM_OXDNA_TO_ANGSTROM
@@ -40,19 +38,19 @@ class Atom(object):
         self.residue_idx = int(pdb_line[22:26])
         self.pos = np.array([float(pdb_line[30:38]), float(pdb_line[38:46]), float(pdb_line[46:54])])
         
-    def is_hydrogen(self):
+    def is_hydrogen(self) -> bool:
         return "H" in self.name
 
     def shift(self, diff:np.ndarray):
         self.pos += diff
 
-    def to_pdb(self, chain_identifier, residue_serial, residue_suffix, bfactor):
-        residue = self.residue + residue_suffix
-        res = "{:6s}{:5d} {:^4s}{:1s}{:3s} {:1s}{:1s}{:4d}{:1s}  {:8.3f}{:8.3f}{:8.3f}{:6.2f}{:6.2f}          {:>2s}{:2s}".format("ATOM", Atom.serial_atom, self.name, " ", residue, chain_identifier," ", residue_serial, " ", self.pos[0], self.pos[1], self.pos[2], 1.00, bfactor, " ", " ", " ")
-        Atom.serial_atom += 1
-        if Atom.serial_atom > 99999:
-            Atom.serial_atom = 1
-        return res
+    def to_pdb(self, residue_suffix:str, bfactor:float) -> Dict:
+        return {
+            "name" : self.name,
+            "residue_name" : self.residue + residue_suffix,
+            "pos" : self.pos,
+            "bfactor" : bfactor
+        }
 
     def to_mgl(self):
         colors = {"C" : "0,1,1", "P" : "1,1,0", "O" : "1,0,0", "H" : "0.5,0.5,0.5", "N" : "0,0,1"}
@@ -61,21 +59,16 @@ class Atom(object):
         r = 0.5
         return "%s %s %s @ %f C[%s]" % (self.pos[0], self.pos[1], self.pos[2], r, color)
 
-class Nucleotide():
-    RNA_warning_printed = False
+class PDB_Nucleotide(object):
     
     def __init__(self, name, idx):
         self.name = name.strip()
         if self.name in NAME_TO_BASE.keys():
             self.base = NAME_TO_BASE[self.name]
         elif self.name in BASES:
-            if self.name == "U" and not Nucleotide.RNA_warning_printed:
-                print("WARNING: unsupported uracil detected: use at your own risk", file=sys.stderr)
-                Nucleotide.RNA_warning_printed = True
-                
             self.base = self.name
         else:
-            self.base = name[1:]
+            self.base = name.strip('D53')
         self.idx = idx
         self.base_atoms = []
         self.phosphate_atoms = []
@@ -120,16 +113,17 @@ class Nucleotide():
             r = self.named_atoms[perm[2]]
             v1 = p.pos - q.pos
             v2 = p.pos - r.pos
-            v1 /= sqrt(np.dot(v1, v1))
-            v2 /= sqrt(np.dot(v2, v2))
-            if abs(np.dot(v1, v2)) > 0.01 or 1:
+            v1 /= np.linalg.norm(v1)
+            v2 /= np.linalg.norm(v2)
+            
+            if abs(np.dot(v1, v2)) > 0.01 or abs(np.dot(v1, v2)) < 1:
                 a3 = np.cross(v1, v2)
-                a3 /= sqrt(np.dot(a3, a3))
+                a3 /= np.linalg.norm(a3)
                 if np.dot(a3, parallel_to) < 0.: 
                     a3 = -a3
                 self.a3 += a3
 
-        self.a3 /= sqrt(np.dot(self.a3, self.a3))
+        self.a3 /= np.linalg.norm(self.a3)
 
     def compute_a1(self):
         if "C" in self.name or "T" in self.name or "U" in self.name:
@@ -144,7 +138,7 @@ class Nucleotide():
             diff = p.pos - q.pos
             self.a1 += diff
 
-        self.a1 /= sqrt(np.dot(self.a1, self.a1))
+        self.a1 /= np.linalg.norm(self.a1)
 
     def compute_as(self):
         self.compute_a1()
@@ -155,8 +149,8 @@ class Nucleotide():
     def correct_for_large_boxes(self, box):
         map(lambda x: x.shift(-np.rint(x.pos / box ) * box), self.atoms)
 
-    def to_pdb(self, chain_identifier, print_H, residue_serial, residue_suffix, residue_type, bfactor):
-        res = []
+    def to_pdb(self, print_H:bool, residue_type:str, bfactor:float) -> List[Dict]:
+        res:List[Dict] = []
         for a in self.atoms:
             if not print_H and 'H' in a.name:
                 continue
@@ -170,7 +164,7 @@ class Nucleotide():
             elif residue_type == "3":
                 if a.name == "O3'":
                     O3prime = a
-            res.append(a.to_pdb(chain_identifier, residue_serial, residue_suffix, bfactor))
+            res.append(a.to_pdb(residue_type, bfactor))
             
         # if the residue is a 3' or 5' end, it requires one more hydrogen linked to the O3' or O5', respectively
         if residue_type == "5":
@@ -179,9 +173,9 @@ class Nucleotide():
             
             # we put the new hydrogen at a distance 1 Angstrom from the O5' oxygen along the direction that, in a regular nucleotide, connects O5' and P
             dist_P_O = phosphorus.pos - O5prime.pos
-            dist_P_O *= 1. / np.sqrt(np.dot(dist_P_O, dist_P_O))
+            dist_P_O /= np.linalg.norm(dist_P_O)
             new_hydrogen.pos = O5prime.pos + dist_P_O
-            res.append(new_hydrogen.to_pdb(chain_identifier, residue_serial, residue_suffix, bfactor))
+            res.append(new_hydrogen.to_pdb(residue_type, bfactor))
         elif residue_type == "3":
             new_hydrogen = copy.deepcopy(O3prime)
             new_hydrogen.name = "HO3'"
@@ -189,11 +183,11 @@ class Nucleotide():
             # we put the new hydrogen at a distance 1 Angstrom from the O3' oxygen along a direction which is a linear combination of the three 
             # orientations that approximately reproduce the crystallographic position
             new_distance = 0.2 * self.a2 - 0.2 * self.a1 - self.a3
-            new_distance *= 1. / np.sqrt(np.dot(new_distance, new_distance))
+            new_distance /= np.linalg.norm(new_distance)
             new_hydrogen.pos = O3prime.pos + new_distance
-            res.append(new_hydrogen.to_pdb(chain_identifier, residue_serial, residue_suffix, bfactor))
+            res.append(new_hydrogen.to_pdb(residue_type, bfactor))
 
-        return "\n".join(res)
+        return res
 
     def to_mgl(self):
         res = []
@@ -224,8 +218,7 @@ class Nucleotide():
 
     atoms = property(get_atoms)
 
-class AminoAcid(object):
-    # RNA_warning_printed = False
+class PDB_AminoAcid(object):
     
     def __init__(self, name, idx):
         object.__init__(self)
@@ -243,7 +236,7 @@ class AminoAcid(object):
         self.chain_id = None
 
     def get_atoms(self):
-        return self.backbone_atoms + self.sidechain_atoms + self.ca_atom
+        return self.backbone_atoms + self.sidechain_atoms + [self.ca_atom]
 
     def add_atom(self, a):
         if a.name in ['N', 'C', 'O']:
@@ -257,83 +250,49 @@ class AminoAcid(object):
             self.chain_id = a.chain_id
 
     def get_ca_pos(self):
-        # Use Residue's Alpha Carbon position
         if self.ca_atom:
             com = np.array(self.ca_atom.pos)
-
+        else:
+            raise RuntimeError("No CA atom found in amino acid residue name:{}, id:{}, chain:{}".format(self.name, self.idx, self.chain_id))
+        
         return com
 
     def get_com(self, atoms=None):
-        if atoms == None: 
-            self.atoms = self.get_atoms()
-            atoms = self.atoms
+        if atoms == None:
+            atoms = self.get_atoms()
         com = np.array([0., 0., 0.])
         for a in atoms:
             com += a.pos
 
         return com / len(atoms)
-
-    # def compute_as(self):
-        # self.compute_a1()
-        # self.compute_a3()
-        # self.a2 = np.cross(self.a3, self.a1)
-        # self.check = abs(np.dot(self.a1, self.a3))
         
     def correct_for_large_boxes(self, box):
         map(lambda x: x.shift(-np.rint(x.pos / box ) * box), self.atoms)
 
-    def to_pdb(self, chain_identifier, print_H, residue_serial, residue_suffix, residue_type):
+    def to_pdb(self, print_H:bool, bfactor:float) -> List[Dict]:
         res = []
-        for a in self.atoms:
-            res.append(a.to_pdb(chain_identifier, residue_serial, residue_suffix))
-            
-        # if the residue is a 3' or 5' end, it requires one more hydrogen linked to the O3' or O5', respectively
-        # if residue_type == "5":
-        #     new_hydrogen = copy.deepcopy(phosphorus)
-        #     new_hydrogen.name = "HO5'"
-            
-        #     # we put the new hydrogen at a distance 1 Angstrom from the O5' oxygen along the direction that, in a regular nucleotide, connects O5' and P
-        #     dist_P_O = phosphorus.pos - O5prime.pos
-        #     dist_P_O *= 1. / np.sqrt(np.dot(dist_P_O, dist_P_O))
-        #     new_hydrogen.pos = O5prime.pos + dist_P_O
-        #     res.append(new_hydrogen.to_pdb(chain_identifier, residue_serial, residue_suffix))
-        # elif residue_type == "3":
-        #     new_hydrogen = copy.deepcopy(O3prime)
-        #     new_hydrogen.name = "HO3'"
-            
-        #     # we put the new hydrogen at a distance 1 Angstrom from the O3' oxygen along a direction which is a linear combination of the three 
-        #     # orientations that approximately reproduce the crystallographic position
-        #     new_distance = 0.2 * self.a2 - 0.2 * self.a1 - self.a3
-        #     new_distance *= 1. / np.sqrt(np.dot(new_distance, new_distance))
-        #     new_hydrogen.pos = O3prime.pos + new_distance
-        #     res.append(new_hydrogen.to_pdb(chain_identifier, residue_serial, residue_suffix))
+        for a in self.get_atoms():
+            if not print_H and 'H' in a.name:
+                continue
+            res.append(a.to_pdb('', bfactor))
 
-        return "\n".join(res)
+        return res
 
-    # def to_mgl(self):
-    #     res = []
-    #     for a in self.atoms:
-    #         res.append(a.to_mgl())
-
-    #     return "\n".join(res)
+    def translate(self, t):
+        for a in self.get_atoms():
+            a.pos += t
 
     def rotate(self, R):
         com = self.get_com()
-        for a in self.atoms:
+        for a in self.get_atoms():
             a.pos = np.dot(R, a.pos - com) + com
-
 
     def set_com(self, new_com):
         com = self.get_com()
-        for a in self.atoms:
-            a.pos += new_com - com - COM_SHIFT * self.a1
+        for a in self.get_atoms():
+            a.pos += new_com - com
 
-    # def set_base(self, new_base_com):
-    #     atoms = [v for k, v in self.named_atoms.iteritems() if k in self.ring_names]
-    #     ring_com = self.get_com(atoms)
-    #     for a in self.atoms:
-    #         a.pos += new_base_com - ring_com - BASE_SHIFT * self.a1
-
-    #     self.compute_as()
-
-    atoms = property(get_atoms)
+    def set_ca_pos(self, new_ca_pos):
+        ca_pos = self.get_ca_pos()
+        for a in self.get_atoms():
+            a.pos += new_ca_pos - ca_pos
