@@ -1,7 +1,7 @@
 import argparse
 import os
 import time
-from typing import List, Tuple
+from typing import List, Tuple, Union
 import numpy as np
 from sys import stderr
 from collections import namedtuple
@@ -15,7 +15,7 @@ ComputeContext = namedtuple("ComputeContext",["traj_info",
                                               "indexes",
                                               "center"])
 
-def svd_align(ref_coords:np.ndarray, coords:np.ndarray, indexes:List[int], ref_center:np.array=np.array([]), center:bool=True) -> Tuple[np.ndarray]:
+def svd_align(ref_coords:np.ndarray, coords:np.ndarray, indexes:List[int], ref_center:np.ndarray=np.array([]), center:bool=True) -> Tuple[np.ndarray, np.ndarray, np.ndarray]:
     """
     Single-value decomposition-based alignment of configurations
 
@@ -26,7 +26,7 @@ def svd_align(ref_coords:np.ndarray, coords:np.ndarray, indexes:List[int], ref_c
         ref_center (numpy.ndarray): (optional) The center of mass of the reference configuration. If not provided, it will be calculated (slightly slower for many confss).
 
     Returns:
-        A tuple of the aligned coordinates (coords, a1s, a3s) for the given chunk
+        Tuple[np.ndarray, np.ndarray, np.ndarray] A tuple of the aligned coordinates (coords, a1s, a3s) for the given chunk
     """
     if len(ref_center) == 0:
         ref_center = np.mean(ref_coords, axis=0)
@@ -62,7 +62,7 @@ def compute(ctx:ComputeContext, chunk_size, chunk_id:int):
     out = ''.join([conf_to_str(c, include_vel=ctx.traj_info.incl_v) for c in confs])
     return out
 
-def align(traj:str, outfile:str, ncpus:int=1, indexes:List[int]=None, ref_conf:Configuration=None, center:bool=True):
+def align(traj:str, outfile:str, ncpus:int=1, indexes:List[int]=[], ref_conf:Union[Configuration,None]=None, center:bool=True):
     """
         Align a trajectory to a ref_conf and print the result to a file.
 
@@ -82,7 +82,7 @@ def align(traj:str, outfile:str, ncpus:int=1, indexes:List[int]=None, ref_conf:C
         #read the first configuration and use it as the reference configuration for the rest
         ref_conf = get_confs(top_info, traj_info, 0, 1)[0]
 
-    if indexes == None:
+    if indexes == []:
         indexes = list(range(top_info.nbases))
 
     ref_conf = inbox(ref_conf) # Don't need to center now because we're going to after indexing anyway.
@@ -147,7 +147,7 @@ def main():
             except:
                 raise RuntimeError("The index file must be a space-seperated list of particles.  These can be generated using oxView by clicking the \"Download Selected Base List\" button")
     else: 
-        indexes = None
+        indexes = []
 
     if args.parallel:
         ncpus = args.parallel[0]
