@@ -24,7 +24,7 @@ ComputeContext_map = namedtuple("ComputeContext_map",["traj_info",
                                                       "centered_ref_coords",
                                                       "components"])
 
-def align_positions(centered_ref_coords:np.ndarray, coords:np.ndarray) -> np.array:
+def align_positions(centered_ref_coords:np.ndarray, coords:np.ndarray) -> np.ndarray:
     """
     Single-value decomposition-based alignment of configurations
 
@@ -70,6 +70,7 @@ def make_heatmap(covariance:np.ndarray):
        xlabel="nucleotide id*3")
     b = fig.colorbar(a, ax=ax)
     b.set_label("covariance", rotation = 270)
+    plt.tight_layout()
     plt.savefig("heatmap.png")
 
 
@@ -145,6 +146,7 @@ def pca(traj_info:TrajInfo, top_info:TopInfo, mean_conf:Configuration, ncpus:int
     plt.scatter(range(0, len(evalues)), evalues, s=25)
     plt.xlabel("component")
     plt.ylabel("eigenvalue")
+    plt.tight_layout()
     plt.savefig("scree.png")
 
     total = sum(evalues)
@@ -159,11 +161,11 @@ def pca(traj_info:TrajInfo, top_info:TopInfo, mean_conf:Configuration, ncpus:int
 
     chunk_size = get_chunk_size()
     coordinates = np.zeros((traj_info.nconfs, ctx.top_info.nbases*3))
-    def callback(i, r):
+    def callback2(i, r):
         nonlocal coordinates, chunk_size
         coordinates[i*chunk_size:(i*chunk_size)+len(r)] = r
 
-    oat_multiprocesser(traj_info.nconfs, ncpus, map_confs_to_pcs, callback, ctx)
+    oat_multiprocesser(traj_info.nconfs, ncpus, map_confs_to_pcs, callback2, ctx)
 
     return (coordinates, evalues, evectors)
 
@@ -213,6 +215,7 @@ def main():
     fig = plt.figure()
     ax = fig.add_subplot(projection='3d')
     ax.scatter(coordinates[:,0], coordinates[:,1], coordinates[:,2], c='g', s=25)
+    plt.tight_layout()
     plt.savefig("coordinates2.png")
 
     #Create an oxView overlays for the first N components
