@@ -13,15 +13,19 @@ struct PSBond {
 	LR_vector force;
 	int p_patch, q_patch;
 	LR_vector r;
+	number r_mod;
 	LR_vector p_torque, q_torque;
+	LR_vector r_part;
 
-	PSBond(BaseParticle *o, number e, number eps, int pp, int qp, LR_vector nr) :
-		other(o),
-		energy(e),
-		epsilon(eps),
-		p_patch(pp),
-		q_patch(qp),
-		r(nr) {
+	PSBond(BaseParticle *o, number e, number eps, int pp, int qp, LR_vector nr, number rm, LR_vector rp) :
+					other(o),
+					energy(e),
+					epsilon(eps),
+					p_patch(pp),
+					q_patch(qp),
+					r(nr),
+					r_mod(rm),
+					r_part(rp) {
 
 	}
 };
@@ -37,23 +41,17 @@ struct PSBondCompare {
 	}
 };
 
-/**
- * @brief Handles interactions in microgel systems.
- *
- * This interaction is selected with
- * interaction_type = CGNucleicAcidsInteraction
- *
- * @verbatim
-
- @endverbatim
- */
 class CGNucleicAcidsInteraction: public BaseInteraction {
 protected:
 	number _Kfene = 15.;
 	number _rfene = 1.5;
 	number _sqr_rfene;
-	number _WCA_sigma = 1.0;
+	number _WCA_sigma = 1.0, _WCA_sigma_unbonded;
 	number _PS_sqr_rep_rcut;
+	number _mu = 1.0;
+	number dS_mod = 1.0;
+	number alpha_mod = 1.0;
+	number bdG_threshold = 1.0;
 
 	std::vector<LR_vector> _chain_coms;
 
@@ -61,6 +59,7 @@ protected:
 	number _PS_beta = 0.;
 	number _PS_gamma = 0.;
 	int _PS_n = 6;
+	int _bead_size = 0;
 
 	int _N_attractive_types = 0;
 	int _interaction_matrix_size = 0;
@@ -79,14 +78,27 @@ protected:
 
 	/// three-body flexibility stuff
 	bool _enable_semiflexibility = false;
+	bool _enable_semiflexibility_3b = false;
+	bool _enable_patch_stacking = false;
 	number _semiflexibility_k;
+	number _semiflexibility_a1;
+	number _semiflexibility_3b_k;
+	number _semiflexibility_3b_exp_sigma = -1.0;
+	number _stacking_eta;
 
 	/// patchy stuff
 	number _deltaPatchMon = 0.5;
 
+	/// used when max_backbone_force = true
+	bool _use_mbf = false;
+	number _mbf_xmax = 0.0;
+	number _mbf_fmax = 0.0;
+	number _mbf_Emax = 0.0;
+	number _mbf_A = 0.0;
+	number _mbf_B = 0.0;
+
 	std::map<int, std::set<PSBond, PSBondCompare> > _bonds;
 
-	int _chain_size = -1;
 	int _N_chains = -1;
 
 	StressTensor _inter_chain_stress_tensor;
@@ -101,6 +113,8 @@ protected:
 	number _sticky(BaseParticle *p, BaseParticle *q, bool update_forces);
 	number _patchy_three_body(BaseParticle *p, PSBond &new_bond, bool update_forces);
 	number _semiflexibility_three_body(BaseParticle *middle, BaseParticle *n1, BaseParticle *n2, bool update_forces);
+	number _semiflexibility_two_body(BaseParticle *p, BaseParticle *q, bool update_forces);
+	number _patch_stacking(BaseParticle *p, BaseParticle *q, bool update_forces);
 
 public:
 	enum {
@@ -120,7 +134,7 @@ public:
 
 	number P_inter_chain();
 
-	virtual void allocate_particles(std::vector<BaseParticle *> &particles);
+	virtual void allocate_particles(std::vector<BaseParticle*> &particles);
 
 	void begin_energy_computation() override;
 	void begin_energy_and_force_computation() override;
@@ -134,10 +148,10 @@ public:
 	virtual number pair_nonbonded_WCA(BaseParticle *p, BaseParticle *q, bool compute_r = true, bool update_forces = false);
 	virtual number pair_nonbonded_sticky(BaseParticle *p, BaseParticle *q, bool compute_r = true, bool update_forces = false);
 
-	virtual void read_topology(int *N_stars, std::vector<BaseParticle *> &particles);
-	virtual void check_input_sanity(std::vector<BaseParticle *> &particles);
+	virtual void read_topology(int *N_stars, std::vector<BaseParticle*> &particles);
+	virtual void check_input_sanity(std::vector<BaseParticle*> &particles);
 };
 
-extern "C" CGNucleicAcidsInteraction *make_CGNucleicAcidsInteraction();
+extern "C" CGNucleicAcidsInteraction* make_CGNucleicAcidsInteraction();
 
 #endif /* CGNUCLEICACIDS_INTERACTION_H */
