@@ -37,7 +37,7 @@ protected:
 	int *_d_number_neighs_no_doubles = nullptr;
 	bool *_d_cell_overflow = nullptr;
 
-	cudaTextureObject_t _counters_cells_tex;
+	cudaTextureObject_t _counters_cells_tex = 0;
 
 	CUDA_kernel_cfg _cells_kernel_cfg;
 
